refactor: Move and group related code
My impression reading through the code in this project is that
top-level statements have gotten a bit intertangled over time;
this tries to reorder things to have a clearer flow and
structure, grouping related things together.

For example, this moves the recently added Error enum up above
all of the API functions, instead of leaving it somewhat
arbitrarily nestled between `cmark_resume_with_options()` and
`cmark_resume_one_event()`.

No code changes were made to the body of functions or types.
This also does not change any of the public APIs, only where
code is located.

* Group padded newline logic at top of text_modifications.rs

* Rename `padding_of()` to `list_item_padding_of()`

* Add 'Public API Functions' header in lib.rs

* Consolidate public enum and struct types at the
  top of lib.rs, instead of leaving them scattered
  amongst the `cmark*()` public functions.

* Consolidate the two `impl State<'_>` statements

* Move `cmark_resume()`, `cmark()` and `cmark_with_options()`
  from the bottom of lib.rs up to the top, before the ~600
  implementation of `cmark_resume_one_event()` that makes up
  the bulk of lib.rs
